Fix: update NFC tag handling and improve APDU command parsing in tests

+++ sdk/src/test/kotlin/com/example/epassport/data/auth/SecureMessagingTest.kt
@@ -20,7 +20,7 @@ class SecureMessagingTest {
     @Test
     fun parseApdu_extendedLeOnly_parsesCorrectly() {
         // [CLA=00, INS=B0, P1=00, P2=00, 0x00, LeHi=0x10, LeLo=0x00] => Le=4096
-        val cmd = byteArrayOf(0x00, 0xB0, 0x00, 0x00, 0x00, 0x10, 0x00)
+        val cmd = byteArrayOf(0x00.toByte(), 0xB0.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x10.toByte(), 0x00.toByte())
         val result = createSm().parseApdu(cmd)
         assertEquals(0, result.lc)
         assertEquals(4096, result.le)
@@ -29,7 +29,7 @@ class SecureMessagingTest {
 
     @Test
     fun parseApdu_extendedLeZero_means65536() {
-        val cmd = byteArrayOf(0x00, 0xB0, 0x00, 0x00, 0x00, 0x00, 0x00)
+        val cmd = byteArrayOf(0x00.toByte(), 0xB0.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte())
         val result = createSm().parseApdu(cmd)
         assertEquals(0, result.lc)
         assertEquals(65536, result.le)
@@ -38,14 +38,14 @@ class SecureMessagingTest {
 
     @Test
     fun parseApdu_extendedLeMax65535() {
-        val cmd = byteArrayOf(0x00, 0xB0, 0x00, 0x00, 0x00, 0xFF.toByte(), 0xFF.toByte())
+        val cmd = byteArrayOf(0x00.toByte(), 0xB0.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0xFF.toByte(), 0xFF.toByte())
         val result = createSm().parseApdu(cmd)
         assertEquals(65535, result.le)
     }
 
     @Test
     fun parseApdu_shortLeOnly_parsesCorrectly() {
-        val cmd = byteArrayOf(0x00, 0xB0, 0x00, 0x00, 0xFF.toByte())
+        val cmd = byteArrayOf(0x00.toByte(), 0xB0.toByte(), 0x00.toByte(), 0x00.toByte(), 0xFF.toByte())
         val result = createSm().parseApdu(cmd)
         assertEquals(0, result.lc)
         assertEquals(255, result.le)
@@ -54,7 +54,7 @@ class SecureMessagingTest {
 
     @Test
     fun parseApdu_shortLcDataWithLe_parsesCorrectly() {
-        val cmd = byteArrayOf(0x00, 0xA4, 0x04, 0x0C, 0x03, 0x01, 0x02, 0x03, 0x00)
+        val cmd = byteArrayOf(0x00.toByte(), 0xA4.toByte(), 0x04.toByte(), 0x0C.toByte(), 0x03, 0x01, 0x02, 0x03, 0x00)
         val result = createSm().parseApdu(cmd)
         assertEquals(3, result.lc)
         assertEquals(0, result.le)
@@ -63,7 +63,7 @@ class SecureMessagingTest {
 
     @Test
     fun parseApdu_shortLcDataWithoutLe_parsesCorrectly() {
-        val cmd = byteArrayOf(0x00, 0xA4, 0x02, 0x0C, 0x02, 0x01, 0x01)
+        val cmd = byteArrayOf(0x00.toByte(), 0xA4.toByte(), 0x02.toByte(), 0x0C.toByte(), 0x02, 0x01, 0x01)
         val result = createSm().parseApdu(cmd)
         assertEquals(2, result.lc)
         assertEquals(-1, result.le)
@@ -73,7 +73,7 @@ class SecureMessagingTest {
     @Test
     fun parseApdu_case5Short_parsesCorrectly() {
         // GET CHALLENGE with Le=8
-        val cmd = byteArrayOf(0x00, 0x84, 0x00, 0x00, 0x08)
+        val cmd = byteArrayOf(0x00.toByte(), 0x84.toByte(), 0x00.toByte(), 0x00.toByte(), 0x08)
         val result = createSm().parseApdu(cmd)
         assertEquals(0, result.lc)
         assertEquals(8, result.le)
@@ -91,7 +91,7 @@ class SecureMessagingTest {
     @Test
     fun parseApdu_le65536_do97EncodedAsZeroZero() {
         // Le=0x0000 → internal le=65536 となる 7バイト Extended APDU
-        val cmd = byteArrayOf(0x00, 0xB0, 0x00, 0x00, 0x00, 0x00, 0x00)
+        val cmd = byteArrayOf(0x00.toByte(), 0xB0.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte())
         val result = createSm().parseApdu(cmd)
         assertEquals(65536, result.le)
         // 65536 は ISO7816-4 上 0x0000 で表現される（0x0100 ではない）
